+++ tests/Engine/FileEngineTest.php
@@ -21,29 +21,17 @@ use Origin\Cache\Exception\Exception;
 
 class FileEngineTest extends \PHPUnit\Framework\TestCase
 {
-    protected $path;
-
     protected function setUp(): void
     {
-        $this->path = sys_get_temp_dir() . '/cache';
-        if (! is_dir($this->path)) {
-            mkdir($this->path);
-        }
-        $cache = new FileEngine(['path' => $this->path, 'duration' => '+ 10 seconds']);
+        $cache = new FileEngine(['duration' => '+ 10 seconds']);
         $cache->clear();
-    }
-
-    public function testInvalidPath()
-    {
-        $this->expectException(InvalidArgumentException::class);
-        new FileEngine(['path' => '/somewhere']);
     }
 
     public function testSet()
     {
-        $cache = new FileEngine(['path' => $this->path]);
+        $cache = new FileEngine(['path']);
         $this->assertTrue($cache->write('foo', 'bar'));
-        $this->assertEquals('bar', unserialize(file_get_contents($this->path . '/origin_foo')));
+        $this->assertEquals('bar', unserialize(file_get_contents(sys_get_temp_dir(). '/cache/origin_foo')));
         $this->assertFalse($cache->write('foo', ''));
     }
     /**
@@ -51,17 +39,18 @@ class FileEngineTest extends \PHPUnit\Framework\TestCase
      */
     public function testGet()
     {
-        $cache = new FileEngine(['path' => $this->path]);
+        $cache = new FileEngine();
         $this->assertFalse($cache->read('foo'));
         $cache->write('foo', 'bar');
         $this->assertEquals('bar', $cache->read('foo'));
     }
+
     /**
      * @depends testSet
      */
     public function testHas()
     {
-        $cache = new FileEngine(['path' => $this->path]);
+        $cache = new FileEngine();
         $this->assertFalse($cache->exists('foo'));
         $cache->write('foo', 'bar');
         $this->assertTrue($cache->exists('foo'));
@@ -71,7 +60,7 @@ class FileEngineTest extends \PHPUnit\Framework\TestCase
      */
     public function testDelete()
     {
-        $cache = new FileEngine(['path' => $this->path]);
+        $cache = new FileEngine();
         $cache->write('foo', 'bar');
         $this->assertTrue($cache->exists('foo'));
         $this->assertTrue($cache->delete('foo'));
@@ -81,7 +70,7 @@ class FileEngineTest extends \PHPUnit\Framework\TestCase
     }
     public function testClear()
     {
-        $cache = new FileEngine(['path' => $this->path]);
+        $cache = new FileEngine();
         $cache->write('foo', 'bar');
         $cache->write('bar', 'foo');
         $this->assertTrue($cache->clear());
@@ -90,20 +79,24 @@ class FileEngineTest extends \PHPUnit\Framework\TestCase
     }
     public function testIncrement()
     {
-        $cache = new FileEngine(['path' => $this->path]);
-        $this->expectException(Exception::class);
-        $cache->increment('counter');
+        $cache = new FileEngine();
+        $this->assertEquals(1, $cache->increment('counter'));
+        $cache->write('counter', 100);
+        $this->assertEquals(101, $cache->increment('counter'));
+        $this->assertEquals(110, $cache->increment('counter', 9));
     }
     public function testDecrement()
     {
-        $cache = new FileEngine(['path' => $this->path]);
-        $this->expectException(Exception::class);
-        $cache->decrement('counter', 9);
+        $cache = new FileEngine();
+        $this->assertEquals(-1, $cache->decrement('counter'));
+        $cache->write('counter', 110);
+        $this->assertEquals(109, $cache->decrement('counter'));
+        $this->assertEquals(100, $cache->decrement('counter', 9));
     }
 
     public function testSetGetDataTypes()
     {
-        $cache = new FileEngine(['path' => $this->path]);
+        $cache = new FileEngine();
         $int = 123;
         $cache->write('int', $int);
         $this->assertEquals($int, $cache->read('int'));