+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.1.1] - 2019-10-31
+### Changed
+- Enabled FileEngine increment / decrement
+- FileEngine will create path if it does not exist
+
 ## [1.1.0] - 2019-10-30
 ### Changed
 - Changed Redis engine to serialize data and work in a unified way with other engines.
@@ -15,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.1] - 2019-10-14
 ### Fixed
-- Bumped configurable verison to 1.x
+- Bumped configurable version to 1.x
 
 ## [1.0.0] - 2019-10-14